requirements.txt
streamlit
pillow
numpy
imageio